requirements.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
Flask==2.0.1
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
